Split out tests into separate files
To make them easier to work with.
